Splitting backend and API routes into individual files, analogous to frontend interface routes
